perf(benchmarks): train torch path in float32 on GPU
The torch runner trained in float64 (`.double()` + float64 tensors) while the
jax and keras backends already use float32 (runner.py, model_builder.py). Double
precision is not required for these tabular models and makes consumer GPUs
(e.g. RTX 5090, ~1/64 float64 throughput) unusable for the Stage-2 search.

Switch the torch train/eval path to float32 and add device placement:
`_torch_device()` selects `$MONONET_TORCH_DEVICE` → CUDA if available → CPU;
model and tensors move to that device; predictions copy back to host. This makes
torch consistent with jax/keras and lets the flavor search run on the GPU.

Test inputs for the torch model builder updated to float32.

+++ benchmarks/_common/model_builder.py
@@ -94,7 +94,7 @@ def _build_torch(cfg: BenchmarkConfig, bundle: DatasetBundle) -> Any:
 
     :param cfg: Benchmark configuration.
     :param bundle: Dataset bundle.
-    :returns: ``nn.Module`` in float64 mode.
+    :returns: ``nn.Module`` in float32 mode.
     """
     import torch
     from torch import nn
@@ -151,7 +151,7 @@ def _build_torch(cfg: BenchmarkConfig, bundle: DatasetBundle) -> Any:
             y = self.head(self.mono_stack(z))
             return torch.sigmoid(y) if binary else y
 
-    return Model().double()
+    return Model()
 
 
 def _build_jax_embed(
